Fixed the Overlay in the SectionWrapper, Moved Counter to Solutions, Fixed Feature Cards
