+++ src/components/About/About.jsx
@@ -5,20 +5,22 @@ import AccordionSection from "./Accordian";
 
 const About = () => {
   return (
+    <>
+    <Counter />
     <SectionWrapper
       images={[
         {
           src: binaries,
-          className: ' ~w-72/96 opacity-90 ',
+          className: ' ~w-72/96 opacity-30 ',
           position:
-            "absolute top-[10%] right-[30%] z-20 overflow-hidden outline-0 outline-green-600",
+            "absolute top-[10%] right-[30%] z-20 outline-0 outline-green-600",
           style: { objectFit: "cover" },
           alt: "Small Ellipse",
         },
         {
           src: smallEllipse,
           className: ' aspect-square ~w-16/48 ',
-          position: "absolute -top-[9%] left-[43%] z-[10]",
+          position: "absolute -top-[9%] left-[43%] -z-[80]",
           alt: "Small Ellipse",
         },
         {
@@ -28,6 +30,7 @@ const About = () => {
         },
       ]}
     >
+    
       <div className='flex flex-col min-h-auto relative z-30 gap-x-6 w-full ~py-12/16 md:~py-4/24 px-8 lg:px-16 xl:px-24 border-0 border-orange-500 '>
         {/* <div>
           <Counter />
@@ -42,6 +45,7 @@ const About = () => {
         </div>
       </div>
     </SectionWrapper>
+    </>
   );
 };
