+++ src/components/Solutions/Solutions.jsx
@@ -4,51 +4,7 @@ import { largeEllipse, loops, smallEllipse, spiralLoop } from "@/assets";
 import { services } from "../../Constants";
 import SectionWrapper from "../shared/SectionWrapper";
 import { GradientButton } from "../shared/GradientButton";
-
-const FeatureCard = ({ item, index }) => {
-  const cardRef = useRef();
-
-  useEffect(() => {
-    const card = cardRef.current;
-    if (!card) return;
-
-    const handleMouseMove = (e) => {
-      const rect = card.getBoundingClientRect();
-      const x = e.clientX - rect.left;
-      const y = e.clientY - rect.top;
-      card.style.setProperty('--mouse-x', `${x}px`);
-      card.style.setProperty('--mouse-y', `${y}px`);
-    };
-
-    card.addEventListener('mousemove', handleMouseMove);
-    return () => card.removeEventListener('mousemove', handleMouseMove);
-  }, []);
-
-  return (
-    <motion.div
-      ref={cardRef}
-      className="relative overflow-hidden bg-white/5 backdrop-blur-lg rounded-lg p-6 text-center hover:bgg-white/10 transition-colors duration-300 group"
-      initial={{ opacity: 0, y: 50 }}
-      animate={{ opacity: 1, y: 0 }}
-      transition={{ duration: 0.5, delay: index * 0.1 }}
-      whileHover={{ scale: 1.05 }}
-    >
-      <div className="absolute inset-0 opacity-0 group-hover:opacity-100 transition-opacity duration-300" />
-      {/*  bg-gradient-to-r from-blue-500/20 to-purple-500/20 */}
-      <div className="absolute inset-0 bg-[radial-gradient(circle_at_var(--mouse-x)_var(--mouse-y),rgba(255,255,255,0.1),transparent_40%)] opacity-0 group-hover:opacity-100 transition-opacity duration-300" />
-      <div className="relative z-10">
-        <motion.img
-          className="w-32 h-32 object-contain mb-6 mx-auto"
-          src={item.img}
-          alt={item.service}
-          whileHover={{ rotate: 5 }}
-        />
-        <h3 className="text-xl font-semibold mb-4 text-white">{item.service}</h3>
-        <GradientButton children={"see more"} className="relative font-semibold px-6 py-2 ~text-sm/lg rounded-full overflow-hiddenn " />
-      </div>
-    </motion.div>
-  );
-};
+import { FeatureCard } from "./FeatureCards";
 
 const Solutions = () => {
   const controls = useAnimation();
@@ -80,13 +36,13 @@ const Solutions = () => {
         },
         {
           src: smallEllipse,
-          className: 'w-1/4 opacity-60',
+          className: 'w-1/10 opacity-60',
           position: 'absolute top-[14%] left-[35%] z-10',
           alt: 'Small Ellipse',
         },
         {
           src: largeEllipse,
-          className: 'w-1/3 opacity-60',
+          className: 'w-1/8 opacity-60',
           position: 'absolute top-[10%] left-[80%] z-10',
           alt: 'Large Ellipse',
         },
@@ -101,7 +57,9 @@ const Solutions = () => {
           hidden: { opacity: 0, y: 50 }
         }}
         transition={{ duration: 0.5, delay: 0.2 }}
-        className="relative container z-30 w-full px-4 sm:px-8 lg:px-16 xl:px-24 font-jakarta py-16 md:py-24"
+        className="relative flex flex-col items-center z-40 md:w-full ~md:~px-12/24 font-jakarta ~py-12/24 ~pb-8/36 border-0 border-orange-500"
+        // md:~py-4/24
+        // px-4 sm:px-8 lg:px-16 xl:px-24
       >
         <motion.div className="text-center mb-12">
           <motion.h3 
@@ -122,7 +80,7 @@ const Solutions = () => {
           </motion.h1>
         </motion.div>
 
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-8">
+        <div className=" grid ~w-sm/screen grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-8 outline-o outline-lime-600">
           {services.map((item, index) => (
             <FeatureCard key={index} item={item} index={index} />
           ))}
