+++ src/index.css
@@ -15,6 +15,8 @@ html {
 
 @media screen and (prefers-reduced-motion: reduce) {
   html {
+    /* position: relative; */
+    /* overflow: hidden; */
     scroll-behavior: auto;
   }
 }
@@ -104,5 +106,6 @@ html {
 
   body {
     @apply bg-background text-foreground;
+    overflow-x: hidden;
   }
 }
