+++ src/lib/ui/magic-card.jsx
@@ -1,4 +1,3 @@
-"use client";;
 import React, { useCallback, useEffect, useRef } from "react";
 import { motion, useMotionTemplate, useMotionValue } from "framer-motion";
